requirements.txt
pyrogram
kucoin-python
python-dotenv